create QuickSort.js
Besides, I put the python code and the java code in 2 folders separately.
