+++ Quicksort Algorithm/JavaScript/QuickSort.js
@@ -1,0 +1,32 @@
+function quickSort(arr, left = 0, right = arr.length - 1) {
+    if (left < right) {
+        const pivotIndex = partition(arr, left, right);
+        quickSort(arr, left, pivotIndex - 1);
+        quickSort(arr, pivotIndex + 1, right);
+    }
+    return arr;
+}
+
+function partition(arr, left, right) {
+    const pivot = arr[right];
+    let i = left - 1;
+    for (let j = left; j < right; j++) {
+        if (arr[j] < pivot) {
+            i++;
+            swap(arr, i, j);
+        }
+    }
+    swap(arr, i + 1, right);
+    return i + 1;
+}
+
+function swap(arr, i, j) {
+    const temp = arr[i];
+    arr[i] = arr[j];
+    arr[j] = temp;
+}
+
+// Example usage:
+const arr = [5, 2, 9, 1, 5, 6];
+const sortedArr = quickSort(arr);
+console.log(sortedArr); // [1, 2, 5, 5, 6, 9]
